Backend/pyproject.toml
[project]
name = "claude-agent-sdk-implementation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "bcrypt>=4.2.0",
    "certifi>=2024.0.0",
    "claude-agent-sdk>=0.1.48",
    "fastapi>=0.135.1",
    "motor>=3.7.1",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.0.0",
    "python-docx>=1.2.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.22",
    "python-pptx>=1.0.2",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.41.0",
    "pydantic[email]>=2.12.5",
]
